+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "robotcode" extension will be documented in this file
 ## [Unreleased]
 
 - update readme
+- semantic tokens now iterate over nodes
 
 ##  0.2.5